Logic: Create abstract class for commands that modify a person - AbstractEditCommand
