+++ src/main/java/seedu/address/logic/parser/CliSyntax.java
@@ -13,5 +13,6 @@ public class CliSyntax {
     public static final Prefix PREFIX_TAG = new Prefix("t/");
     public static final Prefix PREFIX_PUBLIC_ADDRESS = new Prefix("pa/");
     public static final Prefix PREFIX_PUBLIC_ADDRESS_LABEL = new Prefix("l/");
+    public static final Prefix PREFIX_PUBLIC_ADDRESS_NETWORK = new Prefix("c/");
 
 }
